+++ Runtime/Formatting/ICurrencyFormatter.cs
@@ -3,6 +3,6 @@ namespace fefek5.Currency.Runtime
     /// <summary>Turns a balance into text. Picked per currency, or per CurrencyLabel.</summary>
     public interface ICurrencyFormatter
     {
-        string Format(long value);
+        string Format(int value);
     }
 }
